This commit merges several optimization changes back into the working svn sunflower release including:
(1) Texture memory for accel and displ field
(2) __device__ instead of __constant memory
(3) asyncMemcopy to further hide communications, which is paired with similar MPI routines.
    improvements to Kernel_2_impl

There seem to be several functional regressions in the trunk, which
are commented out in this commit. The CPU version, (i.e.,
GPU_MODE=.false.), does not seem to run.